Fix warnings on linux
Linux now compiles (clang++-3.5) without warnings when the following is enabled (these flags are identical to equator mac warnings):
-Wreorder -Wconstant-conversion -Wint-conversion -Woverloaded-virtual -Wuninitialized -Wunused-parameter -Wshorten-64-to-32 -Wstrict-aliasing -Wshadow -Wconversion -Wsign-compare -Werror -Wsign-conversion

+++ modules/juce_audio_basics/buffers/juce_FloatVectorOperations.cpp
@@ -507,7 +507,7 @@ void JUCE_CALLTYPE FloatVectorOperations::clear (float* dest, int num) noexcept
    #if JUCE_USE_VDSP_FRAMEWORK
     vDSP_vclr (dest, 1, (size_t) num);
    #else
-    zeromem (dest, num * sizeof (float));
+    zeromem (dest, (size_t) num * sizeof (float));
    #endif
 }
 
@@ -516,7 +516,7 @@ void JUCE_CALLTYPE FloatVectorOperations::clear (double* dest, int num) noexcept
    #if JUCE_USE_VDSP_FRAMEWORK
     vDSP_vclrD (dest, 1, (size_t) num);
    #else
-    zeromem (dest, num * sizeof (double));
+    zeromem (dest, (size_t) num * sizeof (double));
    #endif
 }
